chore(tasks): create DescriptionInput
